Tried different algorithm combos until mostly working
Negamax functions cleanly now with a few exceptions. When user is "O," the computer's first move often (but not always) is replaced by a zero. It's unclear where/how this occurs. When it does, a tie game is unable to move to the next screen because it doesn't recognize the board as being completely full; because of that rogue zero it isn't.